Add some explicit casts to fix the TESLA build.

+++ tesla/instrumenter/Callee.cpp
@@ -94,7 +94,7 @@ bool TeslaCalleeInstrumenter::runOnModule(Module &M) {
         bool Init = T->RequiresInit();
         bool Clean = T->RequiresCleanup();
 
-        uint32_t Flags =
+        int Flags =
           (Fork ? TESLA_TRANS_FORK : 0)
           | (Init ? TESLA_TRANS_INIT : 0)
           | (Clean ? TESLA_TRANS_CLEANUP : 0)
@@ -103,8 +103,8 @@ bool TeslaCalleeInstrumenter::runOnModule(Module &M) {
         struct tesla_transition Trans = {
           .flags  = Flags,
           .mask   = T->Source().Mask(),
-          .from   = T->Source().ID(),
-          .to     = T->Destination().ID()
+          .from   = (uint32_t)T->Source().ID(),
+          .to     = (uint32_t)T->Destination().ID()
         };
 
         Transitions.push_back(Trans);
